+++ src/main/java/com/westlake/air/pecs/domain/bean/MzIntensityPairs.java
@@ -9,8 +9,8 @@ import lombok.Data;
 @Data
 public class MzIntensityPairs{
 
-    Double[] mzArray;
+    Float[] mzArray;
 
-    Double[] intensityArray;
+    Float[] intensityArray;
 
 }
